Move calendar view to separate module
