+++ calendarview/src/main/java/com/bhaktaprogram/calendarview/EventType.kt
@@ -1,4 +1,4 @@
-package com.bhaktaprogram.main.custom.view
+package com.bhaktaprogram.calendarview
 
 enum class EventType {
     MostImportant,
