+++ calendarview/src/main/res/values/colors.xml
