+++ main/src/main/java/com/bhaktaprogram/main/CalendarFragment.kt
@@ -4,9 +4,9 @@ import android.os.Bundle
 import android.util.Log
 import android.view.View
 import androidx.fragment.app.Fragment
-import com.bhaktaprogram.main.custom.view.CalendarView
-import com.bhaktaprogram.main.custom.view.DayOfMonthUi
-import com.bhaktaprogram.main.custom.view.FakeDaysRepository
+import com.bhaktaprogram.calendarview.CalendarView
+import com.bhaktaprogram.calendarview.DayOfMonthUi
+import com.bhaktaprogram.calendarview.FakeDaysRepository
 
 class CalendarFragment : Fragment(R.layout.calendar_fragment) {
